add new test files

+++ src/main/java/edu/hm/jarzt/Record.java
@@ -21,7 +21,7 @@ public class Record {
         return dataAmount;
     }
 
-    public void setDataAmount(int dataAmount) {
+    public void setDataAmount(long dataAmount) {
         this.dataAmount = dataAmount;
     }
 }
